requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.2.0
pytest==8.4.1
